refactor: перемещает компоненты по папкам
