+++ src/components/RemoveFromFavorites/RemoveFromFavorites.tsx
