+++ src/pages/GroupPage/GroupPage.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from 'react'
 import { useParams } from 'react-router-dom'
-import { ContactCard } from 'src/components/ContactCard'
-import { GroupContactsCard } from 'src/components/GroupContactsCard'
+import { ContactCard } from 'src/components/ContactCard/ContactCard'
+import { GroupContactsCard } from 'src/components/GroupContactsCard/GroupContactsCard'
 import { GroupContactsDto } from 'src/types/dto/GroupContactsDto'
 import { Col, Row, Spinner } from 'react-bootstrap'
 import { useGetContactsQuery } from 'src/redux/contacts'
